# Make fast expression evaluation even faster