@thomaslevesque 2018-11-01
Do you want me to merge as-is, or do you prefer to squash first?