@blairconrad 2018-11-01
Just for fun, I copied the numbers into a Google Sheet and it spat out this:

![image](https://user-images.githubusercontent.com/3275797/47872559-433b2900-dde5-11e8-96e0-54fb7a1bebd4.png)